2023/12/11

+++ src/main/java/creeperbabytea/phlib/common/magic/spellwork/item/wand/WandMaterial.java
@@ -1,12 +1,15 @@
 package creeperbabytea.phlib.common.magic.spellwork.item.wand;
 
+/**
+ * See {@link WandState}
+ */
 public class WandMaterial {
     public static class WandWood {
         private final float stiffness;
         private final float betrayal;
-        private final short sustainability;
+        private final int sustainability;
 
-        public WandWood(float stiffness, float betrayal, short sustainability) {
+        public WandWood(float stiffness, float betrayal, int sustainability) {
             this.stiffness = stiffness;
             this.betrayal = betrayal;
             this.sustainability = sustainability;
@@ -20,7 +23,7 @@ public class WandMaterial {
             return betrayal;
         }
 
-        public short getSustainability() {
+        public int getSustainability() {
             return sustainability;
         }
     }
@@ -28,9 +31,9 @@ public class WandMaterial {
     public static class WandCore {
         private final float strength;
         private final float loyalty;
-        private final short sustainability;
+        private final int sustainability;
 
-        public WandCore(float strength, float loyalty, short sustainability) {
+        public WandCore(float strength, float loyalty, int sustainability) {
             this.strength = strength;
             this.loyalty = loyalty;
             this.sustainability = sustainability;
@@ -44,7 +47,7 @@ public class WandMaterial {
             return loyalty;
         }
 
-        public short getSustainability() {
+        public int getSustainability() {
             return sustainability;
         }
     }
